+++ fmm/src/analysis/c_calling_convention.rs
@@ -1,3 +1,4 @@
+mod call;
 mod context;
 mod error;
 mod function_declaration;
@@ -6,12 +7,7 @@ mod type_;
 
 use self::{context::Context, error::CCallingConventionError};
 use super::{type_check, type_conversion};
-use crate::{
-    build::{self, InstructionBuilder, NameGenerator, TypedExpression},
-    ir::*,
-    types::{self, Type},
-};
-use std::rc::Rc;
+use crate::{ir::*, types::Type};
 
 // TODO Implement the complete C calling convention for all targets.
 //
@@ -36,7 +32,7 @@ pub fn transform(module: &Module, word_bytes: usize) -> Result<Module, CCallingC
             .function_definitions()
             .iter()
             .map(|definition| function_definition::transform(&context, definition))
-            .map(|definition| transform_function_definition(&context, &definition))
+            .map(|definition| call::transform_function_definition(&context, &definition))
             .collect::<Result<_, _>>()?,
     );
 
@@ -48,79 +44,6 @@ pub fn transform(module: &Module, word_bytes: usize) -> Result<Module, CCallingC
     type_check::check(&module)?;
 
     Ok(module)
-}
-
-fn transform_function_definition(
-    context: &Context,
-    definition: &FunctionDefinition,
-) -> Result<FunctionDefinition, CCallingConventionError> {
-    Ok(FunctionDefinition::new(
-        definition.name(),
-        definition.arguments().to_vec(),
-        definition.result_type().clone(),
-        transform_block(context, definition.body())?,
-        definition.options().clone(),
-    ))
-}
-
-fn transform_block(context: &Context, block: &Block) -> Result<Block, CCallingConventionError> {
-    let mut instructions = vec![];
-
-    for instruction in block.instructions() {
-        instructions.extend(transform_instruction(context, instruction)?);
-    }
-
-    Ok(Block::new(
-        instructions,
-        block.terminal_instruction().clone(),
-    ))
-}
-
-fn transform_instruction(
-    context: &Context,
-    instruction: &Instruction,
-) -> Result<Vec<Instruction>, CCallingConventionError> {
-    Ok(match instruction {
-        Instruction::Call(call)
-            if call.type_().calling_convention() == types::CallingConvention::Target =>
-        {
-            match call.function() {
-                // TODO Support complex expressions.
-                Expression::Variable(variable) => {
-                    let builder = InstructionBuilder::new(Rc::new(
-                        NameGenerator::new(format!("{}_c_", call.name())).into(),
-                    ));
-                    let function_type = call.type_();
-                    let pointer = builder.allocate_stack(function_type.result().clone());
-
-                    builder.call(
-                        build::variable(
-                            variable.name(),
-                            type_::transform_function(context, function_type),
-                        ),
-                        call.arguments()
-                            .iter()
-                            .zip(function_type.arguments())
-                            .map(|(argument, type_)| {
-                                TypedExpression::new(argument.clone(), type_.clone())
-                            })
-                            .chain([pointer.clone()])
-                            .collect(),
-                    )?;
-
-                    builder.add_instruction(Load::new(
-                        function_type.result().clone(),
-                        pointer.expression().clone(),
-                        call.name(),
-                    ));
-
-                    builder.into_instructions()
-                }
-                _ => vec![call.clone().into()],
-            }
-        }
-        _ => vec![instruction.clone()],
-    })
 }
 
 #[cfg(test)]
